Added cookies to google requests.
Cookies are retrieved on the first request and stored in ~/.viking/cookies.txt.
The file is recreated (if needed) everytime viking starts.
Currently only google cookies are used. Others may be added if needed.

cookie, user-agent and referer are now added to the request header.

file.c:a_get_viking_dir() was added and should be used for others to query
(and creating if not existed) location of viking directory (~/.viking).

Signed-off-by: Quy Tonthat <qtonthat@gmail.com>