Make viking's default start location somewhere more likely has maps.
The location before this fix was virtually nowhere in the sea which
has no maps from most popular sources. This can be quite confusing
for first time users who don't even know how to navigate to a location
that has maps.

Viking now starts in New York City. Not because it is my most favourite
place on earth (which is Sydney BTW), but it is because we more likely
find maps there.

Signed-off-by: Quy Tonthat <qtonthat@gmail.com>