Fix #2880741: replace gps_open with gps_open_r
In libgpsd 2.90, the logic of gps_open silently changed:
the returned data is no more allocated, but directly points to a static structure.
So, we use gps_open_r instead.

Futhermore, it is better to not mix libc and glib allocator functions.
g_realloc is replaced by native realloc and memory freed by free.