Use combobox to select map type
Since recent changes, it is now possible for the user to add some map types.
So, the list of available map type can become really important.
The previous UI, based on a radio group, became less appropriate than a
combobox menu.
This is particularly true on devices with small screen.