Fix #2946287: open a recent file
this is a HACK to fix opening a recent file.

recent files are stored as a URI. in our case, this is ALWAYS a file:// URI.
the file reading code expects an ordinary filename, so if the filename begins
with file://, this portion of the filename is dropped.

a more proper (and much larger) fix will involve converting file.c to use GFile.

Signed-off-by: Mike Ryan <mikeryan@lacklustre.net>