Auto use of /dev/ttyUSB type device if available on Unix systems.
Assume this is likely to be the GPS device if /dev/ttyUSB (0 or 1) if present (as presumably only there when the device is connected to the system - detected by udev or similar)
User can still use drop down menu / type it in, if the auto one is not the one they want to use.
Otherwise still defaults to /dev/ttyS0 (if that is even available).