Change max cache time to 1 week.
When caching tiles, it's reasonable for the user to be able to make a
tradeoff as to how long a cache lifetime is reasonable.  Even with OSM
mapnik, it's a reasonable choice to accept tiles that are a week old
rather than incurring network latency.  Arguably terraserver tiles (or
any other source that is not expected to change) should be able to be
set to months.