Add track and waypoint highlight for the selected layer or individual item.
Adds VikLayerFuncLayerSelected callback property into the layer, so that action can be taken by the each layer code.
The treeview on selection invokes the layer selection callback. The trwlayer on the selection callback, stores the selected item into the window, and turns highlight on or off as appropriate.
Provided the draw highlight mode is on, the drawing of selected track(s)/ waypoint(s) item(s) is amended, such that a selected track is shown in the highlight colour;
 a selected waypoint has it's text background of the name drawn in the highlight colour or if it has an image associated with it, then that is shown with a highlighted border.
ATM the highlight colour is hard coded to orange.