Add a new select tool at the window level to enable/disbale item selection of the a track or a waypoint.
This may be from any track/waypoint layer upon left mouse clicks in the viewport.
A right mouse click in the viewport brings up the context menu for the track or waypoint selected, which can be used in the normal manner.
Left mouse clicks not near a track or waypoint clears the currently selected item.