Disable autodownloading when dragging the map
I recently discovered that a recent commit introduce a significant
change on CPU usage.
The commit is 71eff775d86be02173e28421cea7f7d3f5a8344. It offers a
smooth pan when drag&drop, but it introduces a huge amount of tile
requests.

For exemple, when I drag the map, I quickly observe around 600
requests on local tile cache while around 10 requests would be enough.

The matter is, in order to keep a smooth dragging, the viewport is
requested to redraw completly. While doing this, all displayed maps
need to check their local cache. And this, for each X event.

This fix is a temporary one.
A better solution implies certainly a significant refactoring
of display/download.

Signed-off-by: Guilhem Bonnefille <guilhem.bonnefille@gmail.com>