Add background thread loading of DEM files.
Although loading of a single DEM file is relatively quick enough for a user not to notice, however loading many DEM files (like say over 50 to cover a country) can easily take sometime - such as the order of 10 seconds (depending on machine speed & number of files). Previously this would lock up the GUI whilst the DEMs are being processed.

This uses the background thread processing feature to load DEM files, with reporting of the number of items to process.