On the Track Properties Window draw an always updating indicator of where the readings are taken from.
Internally this indicator is called a 'blob' (to be different from the marker line indicator) which is drawn in the same function as the track marker (and so is also coloured black), so that the saved image can be updated as necessary. The whole image is now always saved as any part of it could have changed and is restored when the window hasn't changed size.