Fix incorrect cut/copy of sublayers when accessed via the layers panel buttons.
For cut: need to specifically handle the sublayer request, thus needing to add the VikLayerFuncCutItem layer interface and an implementation in the TrackWaypoint layer.

For copy: handler already present, but the layers panel does not call it when a sublayer. The clipboard function already handles layer/sublayer logic, so always call that.