Add ability to move selected trackpoint or waypoint by the mouse within the select tool.
If you want to move the waypoint immediately press shift whilst selecting the waypoint, otherwise one can move it on the second selection.
If you want to move the trackpoint immediately press control whilst selecting the trackpoint, otherwise one can move it on the second selection.

The idea is to prevent accidently moving the item upon initial selection.

Reuse the core of the existing track/waypoint tool move code.