Fix GPX symbol names for Garmin Oregon 450
While testing the GPS export using Garmin Oregon 450 with software
version 4.20, it was noticed that no symbol seemed to be recognized by
the waypoint manager. The problem seems to be that this device
differenciate between lower and upper chars in the symbol names.

Casing of all symbols apart from 'custom placeholder' were modified into
'Title Case'. The hash helper functions had to be changed to ignore the
difference between lower and upper cases. The import had to be extended
to convert lower case names in sym to the version in the table to work
correctly.

Reported-by: Frank Eckelmann <f.eckelmann@gmx.de>
Signed-off-by: Sven Eckelmann <sven@narfation.org>