Add ability to acquire GPS traces stored on OSM
A new menu entry allows to request GPS traces stored on OpenStreetMap.
Currently, only the visible traces are downloaded, ie the traces
contained in the bounding bos of the current viewport.

This feature is really usefull for OSM contributors as it allows to
do some checks before uploading a new trace.

Signed-off-by: Guilhem Bonnefille <guilhem.bonnefille@gmail.com>