factorize gpsbabel command line
The full path of gpsbabel can be searched a single time.
The side effect is you cannot install gpsbabel after viking's start
to solve the issue.

Furthermore, if gpsbabel is missing, it is an error, not just a warning
as something requested by user cannot be done. At the opposite, the
missing of unbuffer command is not a matter as job can be done.

Signed-off-by: Guilhem Bonnefille <guilhem.bonnefille@gmail.com>