Fix SF#3408170: Selected Track Thickness is Always 1px
Allow setting of the highlight line thickness.
Allow track thickness to be read from the TrackWaypoint Layer.
Use this value when setting the highlighted track(s).

NB ATM this only gets done on setting the highlight - thus changing the track thickness properties when already highlighted, won't update the current highlight thickness. This limitation is quite minor, so overall it's acceptable for use.