Add ability to open a TrackWaypoint layer with an external program (default JOSM).
This is ideal for editing around the track and waypoints with OSM, as it's very unlikely Viking will ever be an OSM editor - it's best left to other dedicated programs.
The layer is converted to a temporary GPX file, but it is not deleted as loading it with another program is not instantaneous.
For now it is up to the user to delete it / use system temp cleanup methods.

The program can be changed by the user and is stored in the preferences.

This *hasn't* been tested on Windows OSs. It's possible a full path to the program may be needed to invoke it on such systems.