Make the waypoint name a property of the waypoint.
This is important step in being able to support undo operations.
This also gives the ability to have multiple same named waypoints.

ATM maintain usage of a hash table for the list of waypoints, but instead of using the (forced unique) name for the hash,
use a simple integer as a kind of UUID.