Make the track name a property of the track.
This is important step in being able to support undo operations.
This also gives the ability to have multiple same named tracks.

ATM maintain usage of a hash table for the list of tracks, but instead of using the (forced unique) name for the hash,
use a simple integer as a kind of UUID.