Rework Track Draw by Velocity into Draw by Speed with a simple traffic light colouring scheme.
Determine the colour of the trackpoint (and/or trackline) relative to the average speed
 *  . slow points are red
 *  . average is yellow
 *  . fast points are green

An option is provided to control the percentage factor away from the average speed to determine the transistions to the differing colours.

This now removes the previous Draw by Velocity mode options, which was unclear how it worked and unlikely any one actively uses it.
Thus this means the .vik file format has changed slightly but other than generating warnings all other features will still work.