More intuitive zoom actions involving the mouse to retain the location under the pointer within the viewport bounds.
As per request SF#3547438, which is clearly much better behaviour - since clearly this is the point of interest the user wants to see and thus subsequent zoom requests keep focusing on that area without having to recenter the display.

The modifiers of CTRL+SHIFT and mouse zoom retain keeping the center point in the center (which is in fact a reversal of the previous setting - unlikely to effect many users, as I didn't even know of this feature myself).

Note the keyboard & menu zoom requests always maintain the center point.