Add zoom by bounds capability when using the zoom tool and holding the SHIFT key and left click to select the zoom in area.
Similar to zooming in on the OpenStreetMap default map.

In the zoom tool mode, when the SHIFT key is held down and then left mouse button is pressed it draws a bounding box area on the display.
When the left mouse button is released the zoom is performed on the selected area.