Fix flickering map background when creating a new track and panning the map with the mouse.
Due to the new way of drawing the next potential trackpoint, the pixmap held becomes invalid when the display is panned.
Thus a way of preventing this update is made by exposing the middle button clicks to the tools when they explicitly request it.
Hence the potential trackpoint drawing is temporarily disabled when the map is panning.