Enable option to not write invisible tracks or waypoints in GPX file and don't write them on GPSBabel convert to method.
This is primarily for only transferring enabled items when uploading to a GPS device.
Solidify a_gpx_write_X functions to always specify writing options, although the options can be set to NULL for default behaviour.

Set OSM uploads to not use invisible tracks since the user won't be expecting those to be transferred and probably shouldn't be transferring 'unusual' GPX files with the non-standard "hidden" text anyway