Store track bounding box and only attempt to draw it if some part of it is visible on screen.
Bounds are calculated on TrackWaypoint creation (i.e. loading tracks
from a file or acquiring from an external source).

Bounds are recalculated on trackpoint changes.
Where 'simple' this is done within the vik_track update functions,
however much direct trackpoint list manipulation is performed by the TrackWaypoint layer,
such as the various merge and split functions, so the recalculation is called there.

Thus with multiple tracks, this skips attempting to draw all trackpoints
of every marked visible track when no part of the track is within the
current screen bounds.

This should mean Viking is much faster in drawing.

Note this still attempts to draw all trackpoints of a track should you
zoom 'inside' it. But this is not much of a penalty.