SF#2564997: Enable specific track colours in default draw by track mode.
Colours are allocated automatically (as previously) if no colour has been assigned.
Colour of the track is shown in the treeview as well.
The colour is saved in .vik files.
Changing the colour is available in the track's properties dialog.
By default new routes are red, so tweaked route icons to match.