Restore Google Directions feature
KML is no more supported but gpsbabel support JSON format in HTML page.

Google Directions is no more a separated configure option: there is no
reason to distinguish from other Google feature.

I also disabled the file checking when data is downloaded via
a_babel_convert_from_url function. The reason is this function can be
used to download any sort of data, not only XML. Currently, it is used
to download GPX (for OSM) and HTML (for Google).
An other solution was to select the checking function using the input
type (with a nice serie of tests on strcmp). But as gpsbabel will do the
check, I consider this check is not needed.

Signed-off-by: Guilhem Bonnefille <guilhem.bonnefille@gmail.com>