Add default values for layers.
Add functions to set the default layer parameter data, which is used on initialization.
This uses the GKeyFile mechanism to read and write files.
Using the group type offers a straightforward way to map to the layer name.
Enable GUI access to change the default values.
Rework the layer initialization to use the current default value for each parameter.