Add a zoom selector on status bar
A single click on zoom indicator on status bar opens a zoom selector.
It seems quite natural to click on the indicator to change the corresponding
value. As zoom level is displayed on statusbar, a click on statusbar should
open a zoom selector.

This patch is inspired by ac4478f4c84881436cf9b25a0fa9dc30f889db63
"Add a zoom combo box into the toolbar".

Signed-off-by: Guilhem Bonnefille <guilhem.bonnefille@gmail.com>