Remove zoom combobox in toolbar
Since it is possible to change the zoom level from the status bar,
there is no more need for an other zoom level selector.
Furthermore, it is always better to have a GUI as clean as possible.

This patch remove the combo box allowing to select the zoom level
from the tool bar.

This patch also refactor the zoom levels related list.

Signed-off-by: Guilhem Bonnefille <guilhem.bonnefille@gmail.com>