Fix crashing if creating a DEM layer with no files.
Ensure the layer is realized before accepting updates, as VIK_GTK_WINDOW_FROM_LAYER will fail.

In the DEM case: since there's no files, the thread exits and generates a layer update before the layer is realized.
So it's better not to bother with a thread if there's no files to use.