Position Trackpoint Edit and Waypoint dialogs to not obscure the selected point.
Shift the dialog towards the opposite edge if we think the dialog will be drawn over the selected point.
For Trackpoint Edit this means it gets shifted up or down,
whilst the Waypoint Property/Edit dialog is shifted left or right.

Positioning of the dialog for a new Waypoint is unchanged.