Fix datasource acquiring waypoints from geotag images.
Still need to return something for the 'command'.
I think in some personnel reworking to the acquire.c method I stopped requiring the command to have a value, thus I removed it.
Unfortunately I shouldn't have checked in this change to the geotag datasource, as in the end the acquire rework didn't include this hoped for simplification.