When adding layers insert 'Base' layers (i.e Maps/GeoRef/DEM) below the selected layer.
Consider a common scenario of opening a GPX or KML file so there is now a TrackWaypoint layer.

Then one adds a map layer (previously this then gets inserted above the trackwaypoint layer
and one can no longer see the tracks or waypoints) so now it gets placed below the TRW layer
and hence the GPS info is still visible.

Maintain the ordering of layers when opened from a file.

Update the help file accordingly.