SF#3608264: Improve Statusbar Location Info Maybe Truncated
Reduce precision of location output in DMS to 4 decimal places.
Since (default float output of) 6 dp of precision for one second too detailed be of use!

Reduce space allocated to naming the Tool in use and give it to the location field.
The tool in use (as reported in the statusbar) is not very important so it doesn't matter
does not matter if is truncated here, since the Tool in use is also indicated by:
1. The mouse icon
2. The button depressed on the toolbar