SF#3387591: Improve keyboard copy+paste behaviour when the editing a layer name in the treeview.
Since the default keyboard shortcuts are overridden, we handle them even when a cell is being edited.
Ensure when editing that keyboard shortcuts do something more sensible then copy/pasting layer data.
Thus as we can't get to the internals of the text actually selected,
just copy and paste the layer name.

This also means we can accept pasting in plain text, which can be useful in naming layers.

Also ignore ctrl+<arrow> when editing the treeview which otherwise moves the viewport.