Generalize route finder
This patch generalize the route finder feature. The goal is to allow
the use of different services to compute a route from two clics.

To achieve this I introduced:
- vik_routing_engine as base class for engines
- vik_routing module for registering engines and providing "default"
  features
And I rewrote the Google related feature.

The new design seems quite overkill, but many more feature will be add in
the future. One of them is to rewrite the Acquire -> Google Direction.
An other is to add a "filter" on a route/track in order to allow a finer
route computation, based on more than 2 points.

Signed-off-by: Guilhem Bonnefille <guilhem.bonnefille@gmail.com>