Fix usage of default maps cache directory when it does not end in a separator.
Seems on Windows the preference 'maplayer_default_dir' does not end in a separator.

Thus when using the default value directly in downloading maps, the directory structure is not be created as expected.

Hence ensure for internal usage the cache_dir always ends with a separator.

This also fixes usage when the preferences is manually edited and the map default doesn't end in a separator.