Enable TrackWaypoint sublayer alphabetical sort options.
These are saved in the layer properties for each sublayer types: tracks(including routes) and waypoints.

By default this is defaulted to off in order to preserve potential ordering of named items like:
10
100
1001

Otherwise the alphabetical sort will destory this order and otherwise ATM no other way to restore it.