Add a common method to produce an information message string about a Trackpoint.
Currently for ease of implementation the message format code is a string of characters.
Each character represents what should be inserted in relation to a Trackpoint.
One day it might evolve into something more user friendly with a frontend to control it, perhaps allowing arbitary text too.

However for now at least some control is offered :)