Add function to refine a route with routing service
This patch add a menu entry for route in order to request a
refined version of the route.

The user is now able to plan a journey by creating a simple route
and then request a routing engine to refine this route.
This is a simple solution to extend route finder tool, adding 'via'
points in the journey.

Signed-off-by: Guilhem Bonnefille <guilhem.bonnefille@gmail.com>