Fix crashing when attempting to download maps along a track when multiple maps are enabled.
Detection of 'duplicate' maps is in horrible C - yuck! all this array+pointer code :(
I think it mainly doesn't work as map_layers[] is accessed before being initialized.
Simply bin this stuff as it's hardly useful anyway.
If the user has more than one of the same map - so be it - display them all.