Show Map License information for each map type only once.
Instead of showing the license each time a map type is added,
 save the map types already used into the Viking settings and only display when a new map type is used.
Settings is extended to handle integer lists to achieve this.

Capture changing the map type in the map layer properties which it did not do previously.

Also make the map license information available all of the time from the map layer right click menu.