Enable a history of requested positions.
Positions are stored on specific Go To requests (such entered position coordinates or to waypoint/track positions).
Currently panning and zooming do not modify the requested position.
Moving backwards and forwards through the list is supported sort of like a web browsers history of pages.

As the positions are not specifically named (just raw coordinates),
 nothing advanced like showing the list of positions or jumping over multiple positions is offered.

Finer details such as the maximum size of the history list are controlled by Viking's internal preferences.