Improve the Route Finder Tool
Changed route finder User Interface workings, it is now similar to the new route tool such that it shows the 'as the crow flys' direct line route,
but on creation of each new trackpoint the route from the last trackpoint is calculated using online services.

The user can switch between adding points with the new route tool or with the route finder tool, removed a lot of the vtk->route_finder fields and merged them with the new route fields instead, added code to ensure that double trackpoints get added at the joins between planned route segments.

Changed so that enforcing duplicates in viktrwlayer.c appending a track duplicates the end of the old track not the start of the new

Changed a bit of logic in route finder click to deal with a current empty route not created by the route finder

Like the current Create Track and Route tools, the Route Finder handles clicks when in pan mode and avoids extra dispay sync refreshing to avoid display flicker.

Update the UI during routing to show a busy cursor and info about the requests being made are shown in the statusbar (particularly in case the online requests are slow).

New icons haven been produced for route finder.

Signed-off-by: Rob Norris <rw_norris@hotmail.com>