SF Features#95 - Part 1: Allow and use configuration of an alternative time display.
Add an option consisting of none, UTC, or (Object's) Positional Local Time.

For several places where a time is displayed, show a tooltip of the time in the manner
specified by the alternate time option.

Places with times shown with this tooltip option are:

Trackpoint Properties
Waypoint Properties
Track Properties

The 'Positional Local Time' setting will attempt to work out the timezone at the position of the point (or Track's center)
For the first stage of implementation a simple longitude to hours offset calculation is used (longitude/15).
This will become the fallback when the real lookup implementation is added.