SF Features#95 - Part 2: TimeZone lookup based on an object's position.
For the 'World' time reference frame,
use the k-d tree method (http://en.wikipedia.org/wiki/Kd-tree) to quickly retreive
 the nearest location to the given position.
Then get the timezone associted with that position to use for the time display.

Thus for example a track's time will be displayed in it's timezone rather than the user's current local time.