Enable Map Cache layout control per map.
The default for new users is the standard OSM layout that should be compatible with other tools.

For existing users the legacy method is kept as the default.

Users can change the default used via the layer defaults configuration or on a map by map basis in their .vik files.

A python script tool (run manually) is provided to assist migrating cache from one format to another (see later commits).