A Mapnik Rendering layer
This uses the Mapnik C++ library to render images given an XML configuration file
 (which describes the datasources and how to render them).
The interfacing code is in C++, hence the need for a C++ compiler.
This has been primarily tested with Mapnik 2.2.0 but should work with other versions.

Images are generated at 256x256 pixels in an OSM Tileset manner and are stored directly in the memory map cache.

More complex configurations/datasources (e.g. the main OSM.org Mapnik Stylesheet)
 can make the rendering relatively slow - sometimes several seconds per tile.
For the first iteration this all happens in the main program thread,
 but will be addressed in subsequent commits.

All this can be disabled at the configure stage with '--disable-mapnik'