Reactivate doc subtree
Previously, in order to build the reference documentation only when needed
(read: only some maintainers and only when they want) a solution was created
to deactivate processing of the whole doc subtree. But the side effect was to
not deliver at all the doc subtree if the correct options are not activated.
With an incomplete delivery, some tools like autoreconf (used by packaging
utilities) fail, bringing packager to patch and patch again the sources.

With this change, the doc subtree is no more ignored and the reference
documentation is processed ony when the corresponding ./configure flags is
activated.