SF Bugs#22: Fix changing coordinates when saving and exporting
Use improved double to string conversion to remove extra digits due to
 excessive string length output which previously 'preserved' the imprecision
 of the internal floating point versus the string type number input.

This is the fast and accurate double to string conversion based on Florian Loitsch's Grisu-algorithm.
Written by Andreas Samoljuk under the MIT License.

Thus the numbers saved are now guaranteed to be same read in
 (if the input is already in the shortest string format).

As a bonus this can reduce the size of a typical .vik file (with many points)
 by 5% to 10%.