Improved GPX write to disk reliability on Linux
Often exporting (to a GPX file) will be directly to a GPS Device.
Under Linux it sometimes doesn't actually write it to the disk in a timely manner.
Then one unplugs the device only to subsequently find it hasn't actually got the file.
This is not ideal when you where planning to follow that route/track!
Thus force a disk sync which may help.