SF Bugs#149: [WINDOWS] HTTPS workaround.
Default to skip verfication as it always fails under Windows.
Some incompatibily is suspected in the reliant binary DLLs of libcurl & libnss.
Location & type of certificate stores need to be supported,
but I've not figured it out for the Windows install.

Linux works fine by default.

MAC OS is probably different again, but at least the user can manually set:
 curl_ssl_verifypeer=0
in viking.ini to get it to work if necessary.