Add external GPX layers
Option to store a TRW layer in a GPX file rather than in the .vik file.
This can be specified in the properties dialog of the TRW layer.  Adds
two new properties to the .vik file to specify the layer type and
external file name.

External GPX layers are only loaded if they are visible.  The saves on
file loading times when external files are large.  If a layer is drawn
(made visible) or selected, then it is loaded.  The disadvantage of this
is that summary statistics for parent layers will pot take account of
external layers that are not loaded.

Also adds menu option for open external layer: create a new layer that
opens a GPX file and maintains an external reference to it as an
external layer (rather than importing all the data points into the vik
file).

TRW layers can now be internal, external, or "external no write".  In
the latter, data is loaded from the file but not written back.  E.g. if
you open a GPX recording, you probably won't edit it, but would be
annoyed if you saved your .vik file and the GPX export stripped data
(like GPS speed) from your .gpx file.  Hence if you open a file as an
external layer, it's no write by default.

Also adds right-click menu option to aggregate layers which recurses
through all children and loads all external layers.