Add file save dialog to uibuilder
Instead of the usual file open option, add a new widget for when you
want to specify a file to save to.  This allows file names to be typed
(using GTK's file save action) and checks whether the file exists.  If
it exists it prompts the user whether they want to overwrite.

To implement this the return type of the callback for the vik file entry
functions had to be changed to allow a boolean saying whether to accept
or whether to ask again.

This is introduced for the external layers feature, but committed
separately as it provides general functionality.