Add splicer tool
Used when editing track/routes to change segments of the track/route.
You can first create an alternative segment in the usual way.  Then with
the splicer, select the start and end points of the part of  the
track/route you want to replace.  Then, click on the alternative
segment.  The splicer will swap it in.