Really fix copying TRW layers to ensure all data is copied.
Fix incorrect fix in SHA:3bf99e90315fff1684209ebad99935b0df4ec71a

Of course, the pointer itself still needs to be valid.

Don't stop the unmarshalling of byte data just because a zero is encountered at some locations.
A zero is a valid value in the byte stream (although it might not actually be used).