Change of search to present a list of results
The search dialog is initially similar, but when a search is done, a
list of results is displayed instead of just the first hit visited.

Implemented for Nominatum and Geonames.  Name finder has been removed
because the server has not been available since 2010.  The currently
disabled Google code has been extended only enough to compile, with a
todo left for the candidate parsing code (there is no point in
implemented this for a discontinued API).