SF Features#117: Area calculation tool
Simple rectangular area calculation accessed via the ruler tool and holding down shift key & left mouse.
Then move the mouse around.
Display of the area is shown in the viewport and the statusbar.
The position of the viewport label can be controlled in the advanced preferences,
including the option to turn it off.