Enable right click menu on viewport for Aggregate Layers
Rework menus so that items that have graphical effect on the viewport
(i.e. TAC and Heatmaps) can be used both from the layers panel and the viewport.

Add some simple controls for Track Area Coverage ease of use:
 On/Off, increase/decrease the tile size.
Add new function to determine tracks that contribute to a particular tiles coverage.