Support unreachable tiles concept for Tiles Area Coverage.
This allows marking individual tiles as inaccessible (e.g typically military use only, so no public access) so that the tile is still included for maximum square and clustering count purposes
These tiles are drawn in red on the viewport.

Attempt to load the file unreachable_tiles.txt from the configuration directory,
which contains a list of tiles that will be marked as unreachable.

An example file for the UK is in the data directory.