Ruler drawing: Avoid drawing fictitious line back to upper left corner
Avoid effective division by zero - when the mouse pointer is back at the starting original x,y position.
Since the length goes to zero or very near, we shouldn't attempt to divide a value by it.
Otherwise the result is NaN, which when used as part of the draw_line functions goes awry.