GTK3 Compatibility: Allow Window Resizing
Change final statusbar component from a button into a label and allow the text to be ellipsized.

Otherwise in GTK3 it means the Window can't be resized smaller;
as the width grows to accommodate whatever text is in the Info component,
 which of course can be arbitrarily long.

ATM Still can't resize smaller than the first 5 fixed status fields,
but if you really want a window that small you can hide the statusbar
and then resizing width even smaller is possible.