GTK3: Rework viewport, layer and tool drawing functions
Use updated internal interface or new cairo interface,
 viewport and layers both on to the main cairo context.

Tools use a seperate cairo context which is overlaid on top of the main surface.

Flow configure event through layers, so that layers using a cairo reference need to recreate the reference,
 as the size of the corresponding viewport surface may have changed size. This is skipped for invisible layers.
Configure events are reapplied for invisible layers when they are made visible.