Panel Tabs for General Statistics and Splits
Additional Tabs in the layers panel to show information about the selected track or route.
These are simply the information in the Track Properties dialog.

The tabs can be reordered and shown (or not) as desired.
Splits timing information tab is hidden by default as only of interest to a subset of users.

ATM These tabs don't refresh if a track is changed but there is a simple manual refresh offered.