Make calendar tooltips generation happen only on each calendar update event
Previously the tooltip generation would happen inside the detail() function callback,
which gets called very often by the GTK framework for mouse over events on the calendar.
If there are large number of tracks to analyse then this callback will take a repeatedly relatively long time to process.
So instead work out the relevant tooltips for the month only when the month changes (which is a much less often called event);
and besides we are already doing much of the analysing already when the month changes.

Thus now the detail() function is very fast as it simply returns the already calculated string.