Redraw window at regular intervals not a fixed time after last move
Previously, each redraw would start a timer before the update took
place. Each time a new event came, the timer would be reset. This means
the update can be indefinitely delayed by a long sequence of events.

Instead, only set the timer if it doesn't exist, so an update is
guaranteed to happen at regular intervals, rather than waiting for a
pause in the event stream.

This avoids lags, e.g. when fractional scaling is used on Wayland.