Enable returning the viewport to an already selected Search Result.
Since one might have moved the viewport, clicking on the already selected search result did nothing.
As the entry doesn't change, there is no 'row-changed' event,
 and tree-views don't seem to expose double click events.
Instead manually monitor the button press/release events and see if the actual selection has changed.
If so, then it's on the same entry and redo the viewport location move.
Otherwise, continue on as before (which uses the registered selection function, that does the viewport move).